+++ example/tests/simple-player.entity.spec.ts
@@ -1,4 +1,4 @@
-import { DomainId, HEXColorValueObject, Result, ShortDomainId, TMapper } from "@types-ddd";
+import { BaseDomainEntity, DateValueObject, DomainId, Entity, HEXColorValueObject, Result, ShortDomainId, TMapper, UserNameValueObject } from "@types-ddd";
 import { Player } from "../simple-player.entity";
 
 describe('simple-player.entity', () => {
@@ -148,5 +148,67 @@ describe('simple-player.entity', () => {
 			teamColor: teamColor.value,
 			userId: userId.uid
 		})
+	} )
+	
+	it( 'should validate different types from aggregate', () => {
+
+		interface Props extends BaseDomainEntity {
+			age?: DateValueObject;
+			userId: ShortDomainId;
+			teamColor: HEXColorValueObject;
+			name?: UserNameValueObject;
+		}
+
+		class Example extends Entity<Props> {
+			private constructor (props: Props) {
+				super( props, Example.name );
+			}
+
+			get age(): DateValueObject | undefined {
+			return this.props.age;
+			};
+			get userId(): ShortDomainId{
+				return this.props.userId;
+			};
+			get teamColor(): HEXColorValueObject {
+				return this.props.teamColor;
+			};
+			get name(): UserNameValueObject | null {
+				return this.props.name ?? null;
+			};
+
+			public static create ( props: Props ): Result<Example> {
+				return Result.ok( new Example( props ) );
+			}
+
+		}
+
+		// 	Create value objects
+		const teamColor = HEXColorValueObject.randomColor();
+		const ID = ShortDomainId.create();
+		const userId = DomainId.create();
+		
+		// Create player
+		const player = Example.create( { ID, teamColor, userId, name: null as any } ).getResult();
+		
+		const isNullable = player.checkProps( ['age'] ).isSome( 'null' );
+		const isUndefined = player.checkProps( ['age'] ).isSome( 'undefined' );
+
+		// is undefined
+		expect( player.age ).toBeUndefined();
+		expect( isNullable ).toBeFalsy();
+		expect( isUndefined ).toBeTruthy();
+
+		const nullableName = player.checkProps( ['name'] ).isSome( 'null' );
+		expect( player.name ).toBeNull();
+		expect( nullableName ).toBeTruthy();
+
+		const typesResult = player.checkProps( ['age', 'name'] ).hasSomeTypes( ['null', 'undefined'] );
+		expect( typesResult ).toBeTruthy();
+
+		const player2 = Example.create( { ID, teamColor, userId, age: null as any } ).getResult();
+
+		const typesResult2 = player2.checkProps( ['age'] ).hasSomeTypes( ['null', 'undefined'] );
+		expect( typesResult2 ).toBeTruthy();
 	})
 });